+++ Changelog.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.3] 09-2-2024
+
+### Added 
+
+- ./Tools/PDFParser/PDFParse.py
+- Mothership ruleset to ./Mothership_Data/rules/RulesOutput.md from PDF Parser
+
 ## [1.0.2] 09-27-2024
 
 ### Updated